@@ -1,26 +1,19 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
+import {registerRoute} from 'workbox-routing';
+import {StaleWhileRevalidate} from 'workbox-strategies';
 
-import { registerRoute } from "workbox-routing";
-import { StaleWhileRevalidate } from "workbox-strategies";
-
+// default fn export receiving some useful params
 export default function swCustom(params) {
-  if (params.debug) {
-    console.log("[Docusaurus-PWA][SW]: running swCustom code", params);
-  }
+  const {
+    debug, // :boolean
+    offlineMode, // :boolean
+  } = params;
 
   // Cache responses from external resources
-  registerRoute(
-    (context) =>
-      [
-        /graph\.facebook\.com\/.*\/picture/,
-        /netlify\.com\/img/,
-        /avatars1\.githubusercontent/,
-      ].some((regex) => context.url.href.match(regex)),
-    new StaleWhileRevalidate()
-  );
+  registerRoute((context) => {
+    return [
+      /graph\.facebook\.com\/.*\/picture/,
+      /netlify\.com\/img/,
+      /avatars1\.githubusercontent/,
+    ].some((regex) => context.url.href.match(regex));
+  }, new StaleWhileRevalidate());
 }
